separate tests to usefully units
