+++ tests/DevLiveServer/ChannelAndUserTest.php
@@ -1,22 +1,14 @@
 <?php
 
-namespace PlanetTeamSpeak\TeamSpeak3Framework\Tests\DevServer;
+namespace PlanetTeamSpeak\TeamSpeak3Framework\Tests\DevLiveServer;
 
-use Exception;
 use PHPUnit\Framework\TestCase;
-use PlanetTeamSpeak\TeamSpeak3Framework\Adapter\Adapter;
-use PlanetTeamSpeak\TeamSpeak3Framework\Adapter\ServerQuery;
 use PlanetTeamSpeak\TeamSpeak3Framework\Exception\AdapterException;
 use PlanetTeamSpeak\TeamSpeak3Framework\Exception\HelperException;
 use PlanetTeamSpeak\TeamSpeak3Framework\Exception\ServerQueryException;
-use PlanetTeamSpeak\TeamSpeak3Framework\Exception\TeamSpeak3Exception;
-use PlanetTeamSpeak\TeamSpeak3Framework\Helper\Signal;
-use PlanetTeamSpeak\TeamSpeak3Framework\Node\Host;
-use PlanetTeamSpeak\TeamSpeak3Framework\Node\Node;
-use PlanetTeamSpeak\TeamSpeak3Framework\Node\Server;
 use PlanetTeamSpeak\TeamSpeak3Framework\TeamSpeak3;
 
-class DevLiveServerTest extends TestCase
+class ChannelAndUserTest extends TestCase
 {
     /**
      * ATTENTION
@@ -42,13 +34,8 @@ class DevLiveServerTest extends TestCase
 
     private string $user_test_active;
     private string $ts3_unit_test_userName;
-    private string $ts3_unit_test_signals;
 
     private int $test_cid;
-
-    private int $duration;
-
-    private Server|Adapter|Node|Host $ts3_VirtualServer;
 
     public function setUp(): void
     {
@@ -64,7 +51,6 @@ class DevLiveServerTest extends TestCase
             $this->ts3_unit_test_channel_name = str_replace('DEV_LIVE_SERVER_UNIT_TEST_CHANNEL=', '', preg_replace('#\n(?!\n)#', '', $env[7]));
             $this->user_test_active = str_replace('DEV_LIVE_SERVER_UNIT_TEST_USER_ACTIVE=', '', preg_replace('#\n(?!\n)#', '', $env[8]));
             $this->ts3_unit_test_userName = str_replace('DEV_LIVE_SERVER_UNIT_TEST_USER=', '', preg_replace('#\n(?!\n)#', '', $env[9]));
-            $this->ts3_unit_test_signals = str_replace('DEV_LIVE_SERVER_UNIT_TEST_SIGNALS=', '', preg_replace('#\n(?!\n)#', '', $env[10]));
         } else {
             $this->active = 'false';
         }
@@ -91,44 +77,6 @@ class DevLiveServerTest extends TestCase
      * @throws ServerQueryException
      * @throws HelperException
      */
-    public function test_can_raw_connect()
-    {
-        if ($this->active == 'false') {
-            $this->markTestSkipped('DevLiveServer ist not active');
-        }
-
-        //TODO: infinity connection on connection without access or wrong port
-        $ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri);
-        $nodeInfo = $ts3_VirtualServer->getInfo();
-
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
-        $this->assertEquals('Linux', $nodeInfo['virtualserver_platform']);
-    }
-
-    /**
-     * @throws AdapterException
-     * @throws ServerQueryException
-     * @throws HelperException
-     */
-    public function test_can_ssh_connect()
-    {
-        if ($this->active == 'false') {
-            $this->markTestSkipped('DevLiveServer ist not active');
-        }
-
-        //TODO: infinity connection on connection without access or wrong port
-        $ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri_ssh);
-        $nodeInfo = $ts3_VirtualServer->getInfo();
-
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
-        $this->assertEquals('Linux', $nodeInfo['virtualserver_platform']);
-    }
-
-    /**
-     * @throws AdapterException
-     * @throws ServerQueryException
-     * @throws HelperException
-     */
     public function test_can_get_channel_info()
     {
         if ($this->active == 'false') {
@@ -138,7 +86,7 @@ class DevLiveServerTest extends TestCase
         $ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri);
         $channelInfo = $ts3_VirtualServer->channelGetByName($this->ts3_unit_test_channel_name)->getInfo();
 
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
 
         $this->assertEquals($this->ts3_unit_test_channel_name, $channelInfo['channel_name']);
         $this->assertEquals(1, $channelInfo['channel_flag_permanent']);
@@ -165,7 +113,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals('Play-Test', $cidTest['channel_name']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -204,7 +152,7 @@ class DevLiveServerTest extends TestCase
         }
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -247,7 +195,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals(0, $channelModifiedResult['channel_flag_permanent']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -282,7 +230,7 @@ class DevLiveServerTest extends TestCase
             $this->assertEquals('invalid channelID', $e->getMessage());
         } finally {
             $this->unset_play_test_channel($ts3_VirtualServer);
-            $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+            $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
         }
     }
 
@@ -328,7 +276,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals($testCid4, $result['pid']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -353,7 +301,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals('Standard Channel', $whoamiChannelName['channel_name']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -379,7 +327,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals(75, $channelPermission['i_channel_needed_delete_power']['permvalue']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -407,7 +355,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals(50, $channelPermission['i_channel_needed_subscribe_power']['permvalue']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -439,7 +387,7 @@ class DevLiveServerTest extends TestCase
         $this->assertArrayNotHasKey('i_channel_needed_subscribe_power', $channelPermission);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -462,7 +410,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals($this->ts3_unit_test_userName, $userInfo['client_nickname']);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -493,7 +441,7 @@ class DevLiveServerTest extends TestCase
         $ts3_VirtualServer->clientGetByName($this->ts3_unit_test_userName)->modify(['client_is_talker'=> 0]);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
@@ -518,126 +466,7 @@ class DevLiveServerTest extends TestCase
         $this->assertEquals($userMoved['cid'], $testCid);
 
         $this->unset_play_test_channel($ts3_VirtualServer);
-        $ts3_VirtualServer->getAdapter()->getTransport()->disconnect();
-    }
-
-    /**
-     * @throws AdapterException
-     * @throws ServerQueryException
-     * @throws HelperException
-     * @throws Exception
-     */
-    public function test_ssh_signal_on_wait_timeout()
-    {
-        if ($this->active == 'false' || $this->ts3_unit_test_signals == 'false') {
-            $this->markTestSkipped('DevLiveServer ist not active');
-        }
-
-        //define duration
-        $this->duration = strtotime('+6 minutes');
-
-        try {
-            // Connect to the specified server, authenticate and spawn an object for the virtual server
-            $this->ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri_ssh);
-        } catch(TeamSpeak3Exception $e) {
-            //catch exception
-            echo $e->getMessage();
-        }
-
-        // Register a callback for serverqueryWaitTimeout events
-        Signal::getInstance()->subscribe('serverqueryWaitTimeout', array($this, 'onWaitTimeout'));
-
-        // Register for server events
-        $this->ts3_VirtualServer->serverGetSelected()->notifyRegister('server');
-
-        try {
-            while (true) {
-                $this->ts3_VirtualServer->getParent()->getAdapter()->wait();
-            }
-        }catch(TeamSpeak3Exception $e) {
-            //catch disconnect exception
-            $this->assertEquals("node method 'getTransport()' does not exist", $e->getMessage());
-            $this->assertEquals(0,$e->getCode());
-        }
-    }
-
-    /**
-     * @throws AdapterException
-     * @throws ServerQueryException
-     * @throws HelperException
-     * @throws Exception
-     */
-    public function test_raw_signal_on_wait_timeout()
-    {
-        if ($this->active == 'false' || $this->ts3_unit_test_signals == 'false') {
-            $this->markTestSkipped('DevLiveServer ist not active');
-        }
-
-        //define duration
-        $this->duration = strtotime('+6 minutes');
-
-        try {
-            // Connect to the specified server, authenticate and spawn an object for the virtual server
-            $this->ts3_VirtualServer = TeamSpeak3::factory($this->ts3_server_uri);
-        } catch(TeamSpeak3Exception $e) {
-            //catch exception
-            echo $e->getMessage();
-        }
-
-        // Register a callback for serverqueryWaitTimeout events
-        Signal::getInstance()->subscribe('serverqueryWaitTimeout', array($this, 'onWaitTimeout'));
-
-        // Register for server events
-        $this->ts3_VirtualServer->serverGetSelected()->notifyRegister('server');
-
-        try {
-            while (true) {
-                $this->ts3_VirtualServer->getParent()->getAdapter()->wait();
-            }
-        }catch(TeamSpeak3Exception $e) {
-            //catch disconnect exception
-            $this->assertEquals("node method 'getTransport()' does not exist", $e->getMessage());
-            $this->assertEquals(0,$e->getCode());
-        }
-    }
-
-    /**
-     * @throws AdapterException
-     * @throws ServerQueryException|\PlanetTeamSpeak\TeamSpeak3Framework\Exception\NodeException
-     * @throws Exception
-     */
-    function onWaitTimeout(int $idle_seconds, ServerQuery $serverquery): void
-    {
-        // If the timestamp on the last query is more than 300 seconds (5 minutes) in the past, send 'keepalive'
-        // 'keepalive' command is just server query command 'clientupdate' which does nothing without properties. So nothing changes.
-        if ($serverquery->getQueryLastTimestamp() < time() - 260) {
-            $serverquery->request('clientupdate');
-        }
-
-        // Get data every minute
-        if ($idle_seconds % 60 == 0) {
-            // Resetting lists
-            $this->ts3_VirtualServer->clientListReset();
-            $this->ts3_VirtualServer->serverGroupListReset();
-
-            // Get servergroup client info
-            $this->ts3_VirtualServer->clientList(['client_type' => 0]);
-            $servergrouplist = $this->ts3_VirtualServer->serverGroupList(['type' => 1]);
-
-            $servergroup_clientlist = [];
-            foreach ($servergrouplist as $servergroup) {
-                $servergroup_clientlist[$servergroup->sgid] = count($this->ts3_VirtualServer->serverGroupClientList($servergroup->sgid));
-            }
-
-            // Get virtualserver info
-            $this->ts3_VirtualServer->getInfo(true, true);
-            $this->ts3_VirtualServer->connectionInfo();
-        }
-
-        if (time() >= $this->duration)
-        {
-            $this->ts3_VirtualServer->getParent()->getTransport()->disconnect();
-        }
+        $ts3_VirtualServer->getParent()->getAdapter()->getTransport()->disconnect();
     }
 
     /**
